Added reference date to task groups
Allows assigning a task group at a different time to when it was created.
e.g. a tour in March may have several tasks due at various times leading up to the tour date.
This lets that same group be used for another tour in June with the same tasks and due dates relative to the tour date.