Made tour page saving a fetch request
Should make it less likely for Eilidh to lose work, and make it faster
by not having to rerender every time it's saved. Also should be more
resilient as errors will be properly reported and the server will be
less likely to time out.

Also includes a couple minor fixes that were getting in the way:
- Maps weren't rendering due to an undefined x/y coordinate
- Removed unnecesary console logs from search.js on every click
- Reformatted a couple things