Fix vectordb syncing
Vectordb was trying to insert the same vector multiple times, as the
signal was being called with a seemingly errouneous 'created' argument.
Copied the signal code and adjusted it to make it work, as well as
making it work on the job queue to avoid slowing down the request cycle.